Revert "Feat/events matching (#47)"
This reverts commit f83accb003e08d9a03cd9bbf42c55c6d5d8476d1.
