+++ Dockerfile.dev.client
@@ -8,7 +8,7 @@ ENV REACT_APP_POSTHOG_KEY RxdBl8vjdTwic7xTzoKTdbmeSC1PCzV6sw-x-FKSB-k
 ENV PATH /app/node_modules/.bin:$PATH
 COPY ./packages/client/package.json /app/
 COPY ./package-lock.json /app/
-RUN npm install --legacy-peer-deps
+RUN npm install
 COPY . /app
 RUN npm run format:client
 RUN npm run build:client
