+++ packages/client/cypress/e2e/13-email-with-slack-journey.spec.cy.ts
@@ -3,7 +3,6 @@
 /* eslint-disable @typescript-eslint/no-shadow */
 import credentials from "../fixtures/credentials.json";
 import { loginFunc } from "../test-helpers/loginFunc";
-import setupEventTrigger from "../test-helpers/setupEventTrigger";
 import { tamplatesFunc } from "../test-helpers/templatesFunc";
 
 const { email, password, slackTemplate, userAPIkey, emailTemplate } =
@@ -53,14 +52,19 @@ describe(
       cy.get("#activeJourney").click();
       cy.contains(slackTemplate.name).click();
       cy.get("#exportSelectedTemplate").click();
-      cy.wait(3000);
+
       cy.get("#email").click();
       cy.get("#activeJourney").click();
       cy.contains(emailTemplate.name).click();
       cy.get("#exportSelectedTemplate").click();
 
       cy.get(".react-flow__viewport").get('[data-isprimary="true"]').click();
-      setupEventTrigger(slackTemplate.eventName, slackTemplate.eventName);
+      cy.get("#eventBased").click();
+      cy.contains("Add Condition Or Group").click();
+
+      cy.get('[data-option="events"]').click();
+      cy.get("#events").type(slackTemplate.eventName);
+      cy.get("[data-savetriggerreator] > button").click();
 
       cy.get('[data-isprimary="true"]')
         .get('[data-handlepos="bottom"]')
@@ -85,23 +89,10 @@ describe(
         body: {
           correlationKey: "slackId",
           correlationValue: slackTemplate.slackUid,
-          event: { [slackTemplate.eventName]: slackTemplate.eventName },
+          event: slackTemplate.eventName,
         },
       }).then(({ body }) => {
-        cy.wait(3000);
-        cy.request({
-          method: "POST",
-          headers: {
-            Authorization: `Api-Key ${userAPIkey}`,
-          },
-          url: `${Cypress.env("AxiosURL")}events/job-status/email`,
-          body: {
-            jobId: body[0]?.jobIds?.[0],
-          },
-        }).then(({ body }) => {
-          expect(body).to.equal("completed");
-        });
-
+        cy.wait(1000);
         cy.request({
           method: "POST",
           headers: {
@@ -109,7 +100,7 @@ describe(
           },
           url: `${Cypress.env("AxiosURL")}events/job-status/slack`,
           body: {
-            jobId: body[0]?.jobIds?.[1],
+            jobId: body[0]?.jobIds?.[0],
           },
         }).then(({ body }) => {
           expect(body).to.equal("completed");
