+++ packages/client/cypress/test-helpers/runEmailJourney.ts
@@ -1,5 +1,4 @@
 export default (JourneyName = "Single step flow") => {
-  cy.wait(3000);
   cy.get('[data-disclosure-link="Journey Builder"]').click();
   cy.wait(3000);
   cy.get("button").contains("Create Journey").click();
@@ -18,7 +17,6 @@ export default (JourneyName = "Single step flow") => {
   cy.get("#email").click();
   cy.get("#activeJourney").click();
   cy.contains("TestTemplate").click();
-  cy.wait(1000);
   cy.get("#exportSelectedTemplate").click();
 
   cy.contains("Save").click();
