+++ packages/client/package.json
@@ -21,7 +21,6 @@
     "@types/jest": "27.5.2",
     "@types/node": "16.11.41",
     "@types/react": "18.0.14",
-    "@types/react-autocomplete": "^1.8.6",
     "@types/react-dom": "18.0.5",
     "@wojtekmaj/react-daterange-picker": "^3.4.0",
     "axios": "0.27.2",
@@ -35,7 +34,6 @@
     "posthog-js": "^1.29.3",
     "qs": "6.11.0",
     "react": "18.2.0",
-    "react-autocomplete": "^1.8.1",
     "react-confirm-alert": "^3.0.6",
     "react-dom": "18.2.0",
     "react-flow-renderer": "^10.3.8",
