--- packages/client/src/components/Autocomplete/index.ts
@@ -1,3 +1,0 @@
-import AutoComplete from "./Autocomplete";
-
-export default AutoComplete;
