+++ packages/client/src/components/Elements/Select/Select.tsx
@@ -13,7 +13,7 @@ export interface SelectProps {
     subtitle?: string;
     tooltip?: string;
   }[];
-  id?: string;
+  id: string;
   name?: string;
   autoWidth?: boolean;
   children?: React.ReactNode;
@@ -42,7 +42,6 @@ export interface SelectProps {
   onOpen?: (e: object) => void;
   renderValue?: (value: any) => React.ReactNode;
   disabled?: boolean;
-  wrapperClassnames?: string;
   tick?: boolean;
 }
 
@@ -75,11 +74,10 @@ const Select = (props: SelectProps) => {
     renderValue,
     disabled,
     tick,
-    wrapperClassnames,
   } = props;
   return (
     <>
-      <div id={id} className={`relative ${wrapperClassnames}`}>
+      <div id={id} className="relative">
         <Listbox
           defaultValue={defaultValue}
           disabled={disabled}
