--- packages/client/src/components/TriggerCreater/AndOrSelect.tsx
@@ -1,23 +1,0 @@
-import { Select } from "components/Elements";
-import React, { FC } from "react";
-
-const andOrOptions = [{ value: "and" }, { value: "or" }];
-
-export interface AndOrSelectProps {
-  value: "and" | "or";
-  onChange: (value: "and" | "or") => void;
-  disabled?: boolean;
-}
-
-const AndOrSelect: FC<AndOrSelectProps> = ({ value, onChange, disabled }) => {
-  return (
-    <Select
-      value={value}
-      options={andOrOptions}
-      onChange={onChange}
-      disabled={disabled}
-    />
-  );
-};
-
-export default AndOrSelect;
