+++ packages/server/src/api/customers/customers.service.ts
@@ -26,7 +26,6 @@ import { Queue } from 'bull';
 import { WINSTON_MODULE_NEST_PROVIDER } from 'nest-winston';
 import { createClient } from '@clickhouse/client';
 import { Workflow } from '../workflows/entities/workflow.entity';
-import { attributeConditions } from '@/fixtures/attributeConditions';
 
 export type Correlation = {
   cust: CustomerDocument;
@@ -614,3 +613,55 @@ export class CustomersService {
     );
   }
 }
+
+const attributeConditions = (type: string, isArray: boolean): any[] => {
+  if (isArray) {
+    return [
+      { label: 'contains', id: 'contains', where: '' },
+      { label: 'does not contain', id: 'doesNotContain', where: '' },
+    ];
+  }
+  switch (type) {
+    case 'String':
+      return [
+        { label: 'is equal to', id: 'isEqual', where: '' },
+        { label: 'is not equal to', id: 'isNotEqual', where: '' },
+        { label: 'exists', id: 'exists', where: '' },
+        { label: 'does not exist', id: 'doesNotExist', where: '' },
+        { label: 'contains', id: 'contains', where: '' },
+        { label: 'does not contain', id: 'doesNotContain', where: '' },
+      ];
+    case 'Email':
+      return [
+        { label: 'is equal to', id: 'isEqual', where: '' },
+        { label: 'is not equal to', id: 'isNotEqual', where: '' },
+        { label: 'exists', id: 'exists', where: '' },
+        { label: 'does not exist', id: 'doesNotExist', where: '' },
+        { label: 'contains', id: 'contains', where: '' },
+        { label: 'does not contain', id: 'doesNotContain', where: '' },
+      ];
+    case 'Number':
+      return [
+        { label: 'is equal to', id: 'isEqual', where: '' },
+        { label: 'is not equal to', id: 'isNotEqual', where: '' },
+        { label: 'is greater than', id: 'isGreaterThan', where: '' },
+        { label: 'is less than', id: 'isLessThan', where: '' },
+        { label: 'exists', id: 'exists', where: '' },
+        { label: 'does not exist', id: 'doesNotExist', where: '' },
+      ];
+    case 'Boolean':
+      return [
+        { label: 'is equal to', id: 'isBoolEqual', where: '' },
+        { label: 'is not equal to', id: 'isBoolNotEqual', where: '' },
+      ];
+    case 'Date':
+      return [
+        {
+          label: 'before',
+          id: 'isTimestampBefore',
+          where: '',
+        },
+        { label: 'after', id: 'isTimestampAfter', where: '' },
+      ];
+  }
+};
