+++ packages/server/src/api/events/events.module.ts
@@ -23,8 +23,6 @@ import {
   CustomerKeysSchema,
 } from '../customers/schemas/customer-keys.schema';
 import { AuthModule } from '../auth/auth.module';
-import { Event, EventSchema } from './schemas/event.schema';
-import { EventKeys, EventKeysSchema } from './schemas/event-keys.schema';
 
 @Module({
   imports: [
@@ -39,9 +37,9 @@ import { EventKeys, EventKeysSchema } from './schemas/event-keys.schema';
     ]),
     MongooseModule.forFeature([
       { name: Customer.name, schema: CustomerSchema },
+    ]),
+    MongooseModule.forFeature([
       { name: CustomerKeys.name, schema: CustomerKeysSchema },
-      { name: Event.name, schema: EventSchema },
-      { name: EventKeys.name, schema: EventKeysSchema },
     ]),
     BullModule.registerQueue({
       name: 'email',
@@ -64,6 +62,5 @@ import { EventKeys, EventKeysSchema } from './schemas/event-keys.schema';
     AudiencesService,
     SlackService,
   ],
-  exports: [EventsService],
 })
 export class EventsModule {}
