--- packages/server/src/api/events/schemas/event-keys.schema.ts
@@ -1,18 +1,0 @@
-import { Prop, Schema, SchemaFactory } from '@nestjs/mongoose';
-import { Document } from 'mongoose';
-
-export type EventKeysDocument = EventKeys & Document;
-
-@Schema()
-export class EventKeys {
-  @Prop()
-  key: string;
-
-  @Prop()
-  type: 'String' | 'Number' | 'Boolean' | 'Email';
-
-  @Prop()
-  isArray: boolean;
-}
-
-export const EventKeysSchema = SchemaFactory.createForClass(EventKeys);
