--- packages/server/src/api/events/schemas/event.schema.ts
@@ -1,9 +1,0 @@
-import { Schema, SchemaFactory } from '@nestjs/mongoose';
-import { Document } from 'mongoose';
-
-export type EventDocument = Event & Document;
-
-@Schema({ strict: false })
-export class Event {}
-
-export const EventSchema = SchemaFactory.createForClass(Event);
