+++ packages/server/src/api/workflows/entities/workflow.entity.ts
@@ -6,19 +6,8 @@ export enum TriggerType {
   time_window,
 }
 
-export interface IEventConditions {
-  value: any;
-  attribute?: string;
-  condition?: string;
-  key?: string;
-  type?: string;
-  comparisonType?: string;
-  relationWithNext?: 'and' | 'or';
-  isArray?: boolean;
-}
-
 export class EventProps {
-  conditions: IEventConditions[];
+  event: string;
 }
 
 export class Trigger {
