+++ packages/server/src/api/workflows/workflows.module.ts
@@ -20,7 +20,6 @@ import {
   CustomerKeys,
   CustomerKeysSchema,
 } from '../customers/schemas/customer-keys.schema';
-import { EventKeys, EventKeysSchema } from '../events/schemas/event-keys.schema';
 
 @Module({
   imports: [
@@ -35,7 +34,6 @@ import { EventKeys, EventKeysSchema } from '../events/schemas/event-keys.schema'
     ]),
     MongooseModule.forFeature([
       { name: Customer.name, schema: CustomerSchema },
-      { name: EventKeys.name, schema: EventKeysSchema },
     ]),
     MongooseModule.forFeature([
       { name: CustomerKeys.name, schema: CustomerKeysSchema },
