--- packages/server/src/fixtures/defaultEventKeys.ts
@@ -1,158 +1,0 @@
-export default [
-  {
-    name: '$active_feature_flags',
-    property_type: null,
-  },
-  {
-    name: '$browser',
-    property_type: 'String',
-  },
-  {
-    name: '$browser_version',
-    property_type: 'Number',
-  },
-  {
-    name: '$ce_version',
-    property_type: 'Number',
-  },
-  {
-    name: '$current_url',
-    property_type: 'String',
-  },
-  {
-    name: '$device_id',
-    property_type: 'String',
-  },
-  {
-    name: '$device_type',
-    property_type: 'String',
-  },
-  {
-    name: '$event_type',
-    property_type: 'String',
-  },
-  {
-    name: '$geoip_city_name',
-    property_type: 'String',
-  },
-  {
-    name: '$geoip_continent_code',
-    property_type: 'String',
-  },
-  {
-    name: '$geoip_continent_name',
-    property_type: 'String',
-  },
-  {
-    name: '$geoip_country_code',
-    property_type: 'String',
-  },
-  {
-    name: '$geoip_country_name',
-    property_type: 'String',
-  },
-  {
-    name: '$geoip_latitude',
-    property_type: 'Number',
-  },
-  {
-    name: '$geoip_longitude',
-    property_type: 'Number',
-  },
-  {
-    name: '$geoip_postal_code',
-    property_type: 'Number',
-  },
-  {
-    name: '$geoip_subdivision_1_code',
-    property_type: 'Number',
-  },
-  {
-    name: '$geoip_subdivision_1_name',
-    property_type: 'String',
-  },
-  {
-    name: '$geoip_time_zone',
-    property_type: 'String',
-  },
-  {
-    name: '$host',
-    property_type: 'String',
-  },
-  {
-    name: '$insert_id',
-    property_type: 'String',
-  },
-  {
-    name: '$ip',
-    property_type: 'String',
-  },
-  {
-    name: '$lib',
-    property_type: 'String',
-  },
-  {
-    name: '$lib_version',
-    property_type: 'String',
-  },
-  {
-    name: '$os',
-    property_type: 'String',
-  },
-  {
-    name: '$pathname',
-    property_type: 'String',
-  },
-  {
-    name: '$plugins_deferred',
-    property_type: null,
-  },
-  {
-    name: '$plugins_failed',
-    property_type: null,
-  },
-  {
-    name: '$plugins_succeeded',
-    property_type: null,
-  },
-  {
-    name: '$referrer',
-    property_type: 'String',
-  },
-  {
-    name: '$referring_domain',
-    property_type: 'String',
-  },
-  {
-    name: '$screen_height',
-    property_type: 'Number',
-  },
-  {
-    name: '$screen_width',
-    property_type: 'Number',
-  },
-  {
-    name: '$session_id',
-    property_type: 'String',
-  },
-  {
-    name: '$time',
-    property_type: 'Date',
-  },
-  {
-    name: 'token',
-    property_type: 'String',
-  },
-  {
-    name: '$viewport_height',
-    property_type: 'Number',
-  },
-  {
-    name: '$window_id',
-    property_type: 'String',
-  },
-  {
-    name: '$viewport_width',
-    property_type: 'Number',
-  },
-];
