--- packages/server/src/fixtures/keyTypes.ts
@@ -1,1 +1,0 @@
-export default ['Email', 'String', 'Boolean', 'Number', 'Date'];
