+++ packages/server/src/tracer.ts
@@ -10,14 +10,10 @@ if (process.env.DD_MONITORING) {
         service: p.name,
     });
     (async () => {
-        try {
-            const { data: hostname } = await axios.get(
-                'http://169.254.169.254/latest/meta-data/local-ipv4'
-            );
-            tracer.setUrl(`http://${hostname}:8126`);
-        } catch (e) {
-            console.error(e);
-        }
+        const { data: hostname } = await axios.get(
+            'http://169.254.169.254/latest/meta-data/local-ipv4'
+        );
+        tracer.setUrl(`http://${hostname}:8126`);
     })();
 }
